フォロー・いいね・グループ機能実装+画像エラー解消

+++ app/controllers/public/favorites_controller.rb
@@ -2,17 +2,14 @@ class Public::FavoritesController < ApplicationController
   
   def create
     post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
-    favorite.save
-    redirect_to post_path(post)
+    current_user.favorites.find_or_create_by(post_id: post.id)
+    redirect_back(fallback_location: root_url)
   end
 
   def destroy
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.find_by(post_id: post.id)
-    favorite.destroy
-    redirect_to post_path(post)
+    favorite.destroy if favorite
+    redirect_back(fallback_location: root_url)
   end
-  
-  
 end
